test(ICS): formalising basic ICS testing just text compare
